+++ app.js
@@ -1,16 +1,8 @@
-// crear un servidor
+const express = require('express');
+const app = express();
+const PORT = 3001;
 
-const http = require('http');
-const fs = require('fs');
+app.get('/home', (req, res) => res.send('Server corriendo desde express al home'));
+app.get('/', (req, res) => res.send('pong'));
 
-const server = http.createServer((req, res) => {
-    const file = fs.readFileSync("./public/index.html")
-
-    res.writeHead(200, {
-        "Content-Type": "text/html; charset=UTF-8"
-    });
-
-    res.end(file);
-});
-
-server.listen(3000, () => console.log("Servidor corriendo en puerto 3000"));
+app.listen(PORT, () => console.log(`Escucho el puerto ${PORT}`));
