--- pages/admin/create.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FunkoShop create</title>
-    <link rel="stylesheet" href="../../index.css">
-</head>
-<body>
-    <header>  <!--Inicio de header-->
-        <nav class="navbar container">
-            <picture class="navbar__logo">
-                <a href="../index.html">
-                    <img src="../../image/branding/logo_light_horizontal.svg" alt="Logo de la marca FunkoShop">
-                </a>
-            </picture>
-            <ul class="navbar__menu">
-                <li class="navbar__item">
-                    <a class="navbar__link" href="../../index.html">SALIR</a>
-                </li>
-            </ul>
-        </nav>
-    </header> <!--Fin de header-->
-    <main class="container">
-        <h1 class="create__title">Crear nuevo item</h1>
-        <section class="create__content">
-            <div class="create__select">
-                <label for="category">Categoría</label>
-                <select name="category" id="category" placeholder="Seleccionar">
-                    <option value="Seleccionar" disabled selected hidden>Seleccionar</option>
-                    <option value="Box">Box</option>
-                    <option value="Remeras">Remeras</option>
-                    <option value="Posters">Posters</option>
-                </select>
-                <label for="license">Licencia</label>
-                <select name="license" id="license" placeholder="Seleccionar">
-                    <option value="Seleccionar" disabled selected hidden>Seleccionar</option>
-                    <option value="StarWars">StarWars</option>
-                    <option value="Pokemon">Pokemon</option>
-                    <option value="HarryPotter">Harry Potter</option>
-                </select>
-            </div>
-            <label class="create__product_name" for="product_name">Nombre del producto</label>
-            <input type="text" id="product_name" name="product_name" placeholder="Kakashi Hatake Shippuden Saga">
-            <textarea class="create__textarea" name="product_description" id="product_description" cols="80" rows="8" placeholder="Descripción del producto"></textarea>
-            <div class="create__text">
-                <label for="sku">SKU</label>
-                <input type="text" id="sku" name="sku" placeholder="SSK111AB001">
-                <label for="price">Precio</label>
-                <input type="text" id="price" name="price" placeholder="$ 0.000,00">
-                <label for="stock">Stock</label>
-                <input type="text" id="stock" name="stock" placeholder="0">
-            </div>
-            <div>
-                <label for="discount">Descuento</label>
-                <input type="text" id="discount" name="discount" placeholder="0%">
-                <label for="promo">Cuotas</label>
-                <select name="promo" id="promo" placeholder="3 cuotas sin interés">
-                    <option value="3cuotas" selected>3 cuotas sin interés</option>
-                    <option value="1cuotas">1 pago</option>
-                    <option value="12cuotas">12 cuotas</option>
-                    <option value="18cuotas">18 cuotas</option>
-                </select>
-            </div>
-            <label for="file_image">Imágenes</label>
-            <input type="file" id="file_image" name="file_image">
-            <input type="button" value="Agregar producto">
-            <label for="reset"></label>
-            <input type="reset" value="Limpiar">
-
-
-        </section>
-        
-    </main>
-    <footer class="footer">  <!--Inicio Footer-->
-        <nav class="navbar container">
-            <ul class="navbar__item">
-                <li class="navbar__item">
-                    <a class="navbar__link" href="../../pages/shop/shop.html">SHOP</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="../../pages/admin/register.html">REGISTRARSE</a>
-                </li>
-                <li>
-                    <a class="navbar__link" href="../../pages/admin/login.html">LOGIN</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="../../pages/shop/contacto.html">CONTACTO</a>
-                </li>
-            </ul>
-            <picture>
-                <img src="../../image/branding/isotype.svg" alt="Isotipo de la marca FunkoShop">
-            </picture>
-        </nav>
-        <p class="footer__copy">All rights reserved 2023 - Funkoshop&copy</p>
-    </footer>
-    <script src="https://cdn.jsdelivr.net/npm/iconify-icon@1.0.7/dist/iconify-icon.min.js"></script>
-</body>
-</html>
