--- pages/admin/register.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FunkoShop register</title>
-</head>
-<body>
-    <header>  <!--Inicio de header-->
-        <nav class="navbar container">
-            <picture class="navbar__logo">
-                <a href="./index.html">
-                    <img src="./image/branding/logo_light_horizontal.svg" alt="Logo de la marca FunkoShop">
-                </a>
-            </picture>
-            <ul class="navbar__menu">
-                <li class="navbar__item with-submenu">
-                    <a class="navbar__link with-icon" href="./pages/shop/shop.html">SHOP <iconify-icon icon="majesticons:chevron-down"></iconify-icon></i></a>
-                    <ul class="submenu">
-                        <li class="submenu__item">
-                            <a href="" class="submenu__link">Funko</a>
-                        </li>
-                        <li class="submenu__item">
-                            <a href="" class="submenu__link">Remeras</a>
-                        </li>
-                        <li class="submenu__item">
-                            <a href="" class="submenu__link">Llaveros</a>
-                        </li>
-                    </ul>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="./pages/shop/contacto.html">CONTACTO</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="./pages/admin/login.html>Login">LOGIN</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__cart" href="./pages/shop/cart.html">
-                    <img src="./image/icons/cart-icon.svg" alt="icono de carrito"></a>
-                </li>
-            </ul>
-        </nav>
-    </header> <!--Fin de header-->
-    <main>
-        
-    </main>
-    <footer class="footer">  <!--Inicio Footer-->
-        <nav class="navbar container">
-            <ul class="navbar__item">
-                <li class="navbar__item">
-                    <a class="navbar__link" href="./pages/shop/shop.html">SHOP</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="./pages/admin/register.html">REGISTRARSE</a>
-                </li>
-                <li>
-                    <a class="navbar__link" href="./pages/admin/login.html">LOGIN</a>
-                </li>
-                <li class="navbar__item">
-                    <a class="navbar__link" href="./pages/shop/contacto.html">CONTACTO</a>
-                </li>
-            </ul>
-            <picture>
-                <img src="./image/branding/isotype.svg" alt="Isotipo de la marca FunkoShop">
-            </picture>
-        </nav>
-        <p class="footer__copy">All rights reserved 2023 - Funkoshop&copy</p>
-    </footer>
-    <script src="https://cdn.jsdelivr.net/npm/iconify-icon@1.0.7/dist/iconify-icon.min.js"></script>
-</body>
-</html>
