--- pages/shop/contacto.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Contacto</title>
-</head>
-<body>
-    <header>
-        <p>Contacto</p>
-        <img src="../../image/branding/logo_light_horizontal.svg" alt="Logo FunkoShop">
-        <ul>
-            <li>
-                <a href="../../index.html">Home</a>
-            </li>
-            <li>
-                <a href="../admin/login.html">Login</a>
-            </li>
-            <li>
-                <a href="./pages/shop/contacto.html">Contacto</a>
-            </li>
-            <li>
-                <a href="Shop.html">Shop</a>
-            </li>
-            
-
-        </ul>
-    </header>
-    <main>
-
-    </main>
-</body>
-</html>
