+++ public/index.html
@@ -5,19 +5,19 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>FunkoShop</title>
-    <link rel="stylesheet" href="./css/index.css">
+    <link rel="stylesheet" href="../public/css/index.css">
 </head>
 <body>
     <header>  <!--Inicio de header-->
         <nav class="navbar container">
             <picture class="navbar__logo">
                 <a href="../public/index.html">
-                    <img src="../public/image/branding/logo_light_horizontal.svg" alt="Logo de la marca FunkoShop">
+                    <img src="../public/image/branding/logo_light_horizontal.svg" alt="Logo de la marca FunkoShop horizontal">
                 </a>
             </picture>
             <ul class="navbar__menu">
                 <li class="navbar__item with-submenu">
-                    <a class="navbar__link with-icon" href="./pages/shop/shop.html">SHOP <iconify-icon icon="majesticons:chevron-down"></iconify-icon></i></a>
+                    <a class="navbar__link with-icon" href="../public/pages/shop/shop.html">SHOP <iconify-icon icon="majesticons:chevron-down"></iconify-icon></i></a>
                     <ul class="submenu">
                         <li class="submenu__item">
                             <a href="" class="submenu__link">Funko</a>
